+++ packages/micromanage-cli/lib/service/index.js
@@ -1,0 +1,20 @@
+/*
+ * Copyright IBM Corp. 2022, 2022
+ *
+ * This source code is licensed under the Apache-2.0 license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+import { Command } from 'commander'
+
+import { buildDependenciesCommand } from './dependencies.js'
+import { buildDeployCommand } from './deploy.js'
+
+function buildServiceCommand() {
+  return new Command('service')
+    .configureHelp({ helpWidth: 100 })
+    .description('Commands that operate on services')
+    .addCommand(buildDependenciesCommand())
+    .addCommand(buildDeployCommand())
+}
+
+export { buildServiceCommand }