+++ services/web-app/components/svg-libraries/icon-library/index.js
@@ -4,14 +4,4 @@
  * This source code is licensed under the Apache-2.0 license found in the
  * LICENSE file in the root directory of this source tree.
  */
-
-import LibraryProvider from '../library-provider'
-import Library from './icon-library'
-
-const IconLibrary = ({ site }) => (
-  <LibraryProvider site={site} type="icon">
-    <Library />
-  </LibraryProvider>
-)
-
-export default IconLibrary
+export { default } from './icon-library'